requirements.txt
Django==5.1.1
grpcio
psycopg
chirpstack-api
django-environ